Changing type of nr, A and B to uint32_t from uint64_t.

+++ layout/linear_enum.c
@@ -40,7 +40,7 @@
 
 static const struct c2_layout_enum_ops linear_enum_ops;
 
-int c2_linear_enum_build(uint64_t nr, uint64_t A, uint64_t B,
+int c2_linear_enum_build(uint32_t nr, uint32_t A, uint32_t B,
 			 struct c2_layout_linear_enum **out)
 {
 	struct c2_layout_linear_enum *lin_enum;
